updated_itemView
