+++ mosad_backend/src/main/java/org/rtss/mosad_backend/config/security/WebConfig.java
@@ -1,0 +1,19 @@
+package org.rtss.mosad_backend.config.security;
+
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/api/**") // Apply CORS to specific paths
+                .allowedOrigins("http://localhost:3000") // Allow frontend origin
+                .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS") // Allow specific methods
+                .allowedHeaders("*") // Allow all headers
+                .allowCredentials(true); // Allow cookies if needed
+    }
+}
