+++ mosad_backend/src/main/java/org/rtss/mosad_backend/config/security/WebConfig.java~
@@ -1,0 +1,4 @@
+package org.rtss.mosad_backend.config.security;
+
+public class WebConfig {
+}
