+++ mosad_frontend/src/App.css
@@ -9,10 +9,6 @@ body {
 .app-container {
   max-width: 1200px;
   margin: 0 auto;
-<<<<<<< HEAD
-  padding: 20px;
-=======
   padding: 2rem;
   text-align: center;
->>>>>>> 042afaa020dd8ae63d355745a273bf580b3c2113
 }
