--- mosad_frontend/src/pages/Itemtable.css
@@ -1,118 +1,0 @@
-/* General Styles */
-body {
-  font-family: Arial, sans-serif;
-  background-color: #231f1f6f;
-  color: #2E8B58;
-  margin: 0;
-  padding: 0;
-}
-
-.container {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 20px;
-}
-
-h1 {
-  color: #2E8B58;
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-button {
-  background-color: #2E8B58;
-  color: #FFFFFF;
-  border: none;
-  padding: 10px 20px;
-  font-size: 16px;
-  cursor: pointer;
-  border-radius: 5px;
-  transition: background-color 0.3s ease;
-}
-
-button:hover {
-  background-color: #FFC220;
-}
-
-/* Table Styles */
-table {
-  width: 100%;
-  border-collapse: collapse;
-  margin-bottom: 20px;
-  background-color: #FFFFFF;
-  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-}
-
-table th, table td {
-  padding: 15px;
-  text-align: left;
-  border: 1px solid #B5FCB5;
-}
-
-table th {
-  background-color: #2E8B58;
-  color: #FFFFFF;
-}
-
-/* Modal Styles */
-.react-modal {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: #FFFFFF;
-  padding: 20px;
-  border-radius: 8px;
-  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.2);
-  width: 100%;
-  max-width: 1400px;
-  z-index: 1000;
-}
-
-.modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 999;
-}
-
-/* Button Alignment */
-.add-item-button {
-  display: flex;
-  justify-content: flex-end;
-  margin-bottom: 20px;
-}
-#itemdetails,#pricedetails{
-  border:1px solid #FFFFFF;
-  margin: 5px;
-  border-radius: 10px;
-  background-color: #ffffff;
-
-}
-
-#pricedetails{
-  
-  color: #000000;
-  input{
-    max-width: 200px;
-  }
-}   
-
-
-
-
-/* Responsive Styles */
-@media (max-width: 768px) {
-  table th, table td {
-    padding: 10px;
-    font-size: 14px;
-  }
-
-  button {
-    font-size: 14px;
-    padding: 8px 16px;
-  }
-}
